+++ src/styles/components/footer.css
@@ -1,0 +1,54 @@
+/** FORMIDABLE Footer
+* A <Footer> component retaining colors, fonts, and space-theme associated with
+* the Formidable brand.
+*
+* -logo         wrapper for Formidable logo
+* -logo-link    link containing an svg of Formidable Logo to formidable.com
+* .default      wrapper for Formidable-specific links; by default: Contact Careers Twitter GitHub
+* .default:last-child   wrapper for trademark prop
+*/
+.formidableFooter {
+  margin: 2rem 0 0 0;
+}
+
+.formidableFooter-logo {
+  height: 50px;
+  margin: 0 0 0 auto;
+  overflow: hidden;
+}
+
+.formidableFooter-logo-link {
+  display: flex;
+  height: inherit;
+}
+
+.formidableFooter .default {
+  margin-top: 16px; /* Align baseline of logo with baseline of link text */
+  font-family: "akkurat", "Inconsolata", monospace;
+  font-size: 13px;
+}
+
+.formidableFooter .default a {
+  display: inline-block;
+  letter-spacing: 0.15em;
+  line-height: 2;
+  margin-right: 2em;
+  text-transform: uppercase;
+}
+
+.formidableFooter .default a:last-child {
+  margin-right: 0;
+}
+
+.formidableFooter .default:last-child { /* Target the trademark */
+  flex: 0 0 100%;
+  margin-top: 8px;
+  text-align: right;
+}
+
+/** DARK THEME
+*/
+.formidableFooter.isDark {
+  background: #242121 linear-gradient(to top, rgba(10,9,9,0) 85%, rgba(10,9,9,0.75) 100%);
+  color: #d1cecd;
+}
